move displaced test
